Fix: don't crash when encountering garbage diffs
Issue tries to decode issue diffs as valid JSON data and would crash
when that assumption failed during tag gathering. Now it will just
report the problem and let the user deal with it.

How to fix it? In my case the problem was empty diffs (how did that
happen?). Removing them fixed the problem.