Address Github issue #12: editing issue descriptions
It is not something that I had a need for (I just added comments if I
needed details to be added or amended), but I can imagine that it can
come in handy -- as evidenced by the question on Github.

So here it is: a quick and dirty solution to the problem.

HOW TO EDIT AN ISSUE?

    ]$ issue open -E <existing-id> [<message>]